Support acmart abstract
